bugfix. do animation in nested CA transaction.
In animation mode, the control was using [CATransaction commit] but
without a matching a [CATransaction begin]. As a result, it was forcing
an early commit of the default transaction context, rather than
containing the control's animations within a nested context.

This could have the effect of modifying other animation that were being
collected into that context, from CALayers unrelated to this control.